+++ lib/features/home/data/datasources/home_remote_datasource.dart
@@ -1,0 +1,8 @@
+abstract class HomeRemoteDataSource {
+  // TODO: Add datasource methods
+}
+
+class HomeRemoteDataSourceImpl implements HomeRemoteDataSource {
+  // final HttpClient client;
+  // HomeRemoteDataSourceImpl(this.client);
+}