+++ lib/features/home/domain/repositories/home_repository.dart
@@ -1,0 +1,6 @@
+import 'package:dartz/dartz.dart';
+import '../entities/home_entity.dart';
+
+abstract class HomeRepository {
+  // TODO: Add repository methods
+}