+++ lib/features/home/domain/usecases/updateprofile_usecase.dart
@@ -1,0 +1,12 @@
+import 'package:dartz/dartz.dart';
+import '../repositories/home_repository.dart';
+
+class UpdateprofileUseCase {
+  final HomeRepository repository;
+
+  UpdateprofileUseCase(this.repository);
+
+  // Future<Either<Failure, dynamic>> call() async {
+  //   return await repository.execute();
+  // }
+}