+++ lib/templates/cubit_template.dart
@@ -7,11 +7,12 @@ class CubitTemplate {
     return '''
 import 'package:flutter_bloc/flutter_bloc.dart';
 import '${name}_state.dart';
+import '../../domain/usecases/get_${name}_usecase.dart';
 
 class ${pascal}Cubit extends Cubit<${pascal}State> {
-  ${pascal}Cubit() : super(const ${pascal}State());
+  final Get${pascal}UseCase get${pascal}UseCase;
 
-
+  ${pascal}Cubit({required this.get${pascal}UseCase}) : super(const ${pascal}State());
 }
 ''';
   }